especificando bigInt como int

+++ src/controllers/rankingController.js
@@ -3,8 +3,8 @@ import { connection } from '../databases/postgres.js';
 export async function getRanking(req, res) {
   try {
     const { rows: ranking } = await connection.query(
-      `SELECT users.id, users.name, COUNT(urls.url_link) AS "linksCount", 
-      COALESCE(SUM(urls.visit_count),0 ) AS "visitCount"
+      `SELECT users.id, users.name, COUNT(urls.url_link)::int AS "linksCount", 
+      COALESCE(SUM(urls.visit_count),0 )::int AS "visitCount"
       FROM users LEFT JOIN urls ON urls.user_id = users.id 
       GROUP BY users.id 
       ORDER BY "visitCount" DESC
